+++ QingToolbox.Shell/App.xaml.cs
@@ -262,7 +262,9 @@ public partial class App : Application
             {
                 services.AddSingleton<WebShellState>();
                 services.AddSingleton<WebNavigationPolicy>();
-                services.AddSingleton(_ => new WebAssetIdentity(Path.Combine(AppContext.BaseDirectory, "WebUI")));
+                services.AddSingleton(_ => new Lazy<WebAssetIdentity>(() =>
+                    new WebAssetIdentity(Path.Combine(AppContext.BaseDirectory, "WebUI")), true));
+                services.AddSingleton<WebActivationSession>();
                 services.AddSingleton<WebAppSnapshotProvider>();
                 services.AddSingleton<IWebCommandHandler, WebPingCommandHandler>();
                 services.AddSingleton<IWebCommandHandler, WebSnapshotCommandHandler>();